Renamed cli command
