+++ cli/utils/prepare-library-package.js
