+++ config/webpack/build-public-library.webpack.config.js
