+++ test/cli-utils-prepare-library-package.spec.js
@@ -1,8 +1,8 @@
 /*jshint jasmine: true, node: true */
 'use strict';
 
-describe('cli utils prepare-package', () => {
-  const util = require('../cli/utils/stage-ts');
+describe('cli utils prepare-library-package', () => {
+  const util = require('../cli/utils/prepare-library-package');
 
   it('should', () => {
     expect(util).toBeDefined();
