+++ test/cli-utils-stage-library-ts.spec.js
@@ -1,8 +1,8 @@
 /*jshint jasmine: true, node: true */
 'use strict';
 
-describe('cli utils prepare-package', () => {
-  const util = require('../cli/utils/prepare-package');
+describe('cli utils prepare-library-package', () => {
+  const util = require('../cli/utils/stage-library-ts');
 
   it('should', () => {
     expect(util).toBeDefined();
