+++ test/config-webpack-build-public-library.spec.js
@@ -1,11 +1,11 @@
 /*jshint jasmine: true, node: true */
 'use strict';
 
-describe('config webpack release', () => {
+describe('config webpack build public library', () => {
   beforeEach(() => {});
 
   it('should expose a getWebpackConfig method', () => {
-    const lib = require('../config/webpack/release.webpack.config');
+    const lib = require('../config/webpack/build-public-library.webpack.config');
     expect(typeof lib.getWebpackConfig).toEqual('function');
   });
 });
